% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funpark_data.R
\docType{data}
\name{alphabets}
\alias{alphabets}
\title{alphabets}
\format{A data.frame with 26 rows and 2 variables:
\describe{
    \item{upper}{<chr> uppercase letters}
    \item{lower}{<chr> lowercase letters}
}}
\usage{
data(alphabets)
}
\description{
Collection of all English alphabets, both lowercases and uppercases.
}
\examples{
data(alphabets)
all_alphabets <- sort(c(alphabets$upper, alphabets$lower))
}
\keyword{datasets}
