% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{binarize_binom}
\alias{binarize_binom}
\title{binarize_binom}
\usage{
binarize_binom(dat, responses, variable.name = NULL)
}
\arguments{
\item{dat}{A data whose features are categorical, and has two integer
columns that correspond to binomial responses}

\item{responses}{A character vector of length 2 where each element is the
names of columns that store the counts of positive and negative
responses, in this order}

\item{variable.name}{NULL by default; a character that will be used for
the name of new binary response column after binarization; if NULL,
then a randomly generated character will be used}
}
\value{
A data.table that has the same categorical features as \code{dat},
  and has \code{variable.name} column as its binary response
}
\description{
Transform a data with binomial responses into a data with binary response.
}
\examples{
# data(rich_binom)
binarize_binom(rich_binom, c('rich', 'not_rich'))
binarize_binom(rich_binom, c('rich', 'not_rich'), 'is_rich')
}
\seealso{
\code{\link{binarize_pois}},
  \code{\link{change_form}},
  \href{https://joon3216.github.io/research_materials/2018/binarize}{Binarizing data using data.table}
}
\author{
Junkyu Park
}
