% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{binarize_pois}
\alias{binarize_pois}
\title{binarize_pois}
\usage{
binarize_pois(dat, response)
}
\arguments{
\item{dat}{A data whose features are categorical, and has a column of
natural numbers that correspond to the count response}

\item{response}{A character; the name of column that stores the count
response}
}
\value{
A data.table that has all of categorical features in \code{dat},
  where each given combination of features (rows) are populated by integers
  based on \code{response}. The function will not keep those combinations
  with a zero count.
}
\description{
Transform a data with a count response into a data with binary response.
}
\examples{
# data(rich_pois)
binarize_pois(rich_pois, 'count')
}
\seealso{
\code{\link{binarize_binom}}, \code{\link{change_form}}
}
\author{
Junkyu Park
}
