% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{change_form}
\alias{change_form}
\title{change_form}
\usage{
change_form(dat, from, to,
            old_response, category, new_response)
}
\arguments{
\item{dat}{A data whose features are categorical and response(s) is/are
nonnegative integer vector(s).}

\item{from}{A character; either 'binary', 'binomial', or 'poisson'}

\item{to}{A character != from; either 'binary', 'binomial', or 'poisson'}

\item{old_response}{(always specified) a character vector of:
\itemize{
  \item length 1 if from = 'binary' or 'poisson'; the name of column
in dat that stores a response/count.
  \item lenght 2 if from = 'binomial'; the names of columns in dat
that store positive and negative case counts, in this order.
}}

\item{category}{(specified only if to = 'poisson') a character vector of:
\itemize{
  \item length 1 if from = 'binomial'; the new name of column that
will store two names in old_response as positive and
negative cases, in this order.
  \item length 2 if from = 'binary'; the new names for positive and
negative cases in the binary response column, in this order.
}}

\item{new_response}{(specified only if from != 'poisson') a character
vector of:
\itemize{
  \item length 1 if to = 'binary' or 'poisson'; the name of the new
column in new data that will store either a binary or count
response.
  \item length 2 if to = 'binomial'; the names of two columns in
new data that will store positive and negative case counts,
in this order.
}}
}
\value{
A data.table that has transformed from the form specified in
  \code{from} to the form specified in \code{to}.
}
\description{
Transform a data from one form into another.
}
\examples{
change_form(
    rich,
    from = 'binary', to = 'binomial',
    old_response = 'is_rich',
    new_response = c('rich', 'not_rich')
)
change_form(
    rich,
    from = 'binary', to = 'poisson',
    old_response = 'is_rich',
    category = c('rich', 'not_rich'),
    new_response = 'count'
)
change_form(
    rich_binom,
    from = 'binomial', to = 'binary',
    old_response = c('rich', 'not_rich'),
    new_response = 'is_rich'
)
change_form(
    rich_binom,
    from = 'binomial', to = 'poisson',
    old_response = c('rich', 'not_rich'),
    category = 'is_rich',
    new_response = 'count'
)
change_form(
    rich_pois,
    from = 'poisson', to = 'binary',
    old_response = 'count'
)
change_form(
    rich_pois,
    from = 'poisson', to = 'binomial',
    old_response = 'count',
    category = 'is_rich'
)
}
\seealso{
\code{\link{binarize_binom}},
  \code{\link{binarize_pois}},
  \href{https://joon3216.github.io/research_materials/2018/binarize}{Binarizing data using data.table}
}
\author{
Junkyu Park
}
