% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{CI_auc}
\alias{CI_auc}
\title{CI_auc}
\usage{
CI_auc(dat, fmlr, R = 500, type = 'norm', ...)
}
\arguments{
\item{dat}{A data whose response is binary (0 and 1).}

\item{fmlr}{A formula for the logistic regression with logit link.}

\item{R}{The same as \code{R} in boot::boot.}

\item{type}{The same as \code{type} in boot::boot.ci}

\item{\dots}{Additional arguments of boot::boot.ci}
}
\value{
A \code{bootci} object.
}
\description{
Calculate a confidence interval of AUC by bootstrapping.
}
\examples{
# library(data.table)
data(femsmoke, package = 'faraway')
femsmoke_binary <- binarize_pois(femsmoke, 'y')
femsmoke_binary[, dead := ifelse(dead == 'yes', 1, 0)]
CI_auc(femsmoke_binary, dead ~ smoker + age)
}
\seealso{
\code{\link{binarize_binom}},
  \code{\link{binarize_pois}},
  \code{\link{change_form}},
  \code{\link{plot_roc}},
  \href{https://joon3216.github.io/research_materials/2018/binarize}{Binarizing data using data.table}
}
\author{
Junkyu Park
}
