% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_separable_penalty.R
\name{fusion_estimates}
\alias{fusion_estimates}
\title{fusion_estimates}
\usage{
fusion_estimates(y, theta, lambda, max_iter = 100, eps = 1e-5)
}
\arguments{
\item{y}{A vector of numbers.}

\item{theta}{A vector of numbers that has the same length as \code{y}. If
missing, then it is replaced with \code{y}.}

\item{lambda}{A positive number; the tuning parameter.}

\item{max_iter}{A natural number; the maximum iteration. \code{100} by
default.}

\item{eps}{A positive number; \code{1e-05} by default.}
}
\value{
A list:
\itemize{
  \item \code{theta}: Resulting fusion estimates.
  \item \code{phi}: Differences in \code{theta}.
  \item \code{lambda}: A specified \code{lambda} when running the function.
  \item \code{iter}: The number of iterations.
  \item \code{costs}: Collection of costs at each iteration.
}
}
\description{
Find the fusion estimates of given \code{y}. Start iteration with \code{theta}
and the tuning parameter \code{lambda}.
}
\examples{
set.seed(1024)
n <- 1000
t <- 1:n
f <- function(t) {t / 250 - .5}
g <- function(t) {-(.25 / 449) * t + 250 / 449}
true_theta <- c(rep(0, 249), f(250:500), rep(.75, 50), g(551:1000))
y <- true_theta + rnorm(1000, 0, 0.1)
start <- rep(.5, n)
result <- fusion_estimates(y, theta = Start, lambda = 1, max_iter = 1000)
}
\seealso{
\href{https://joon3216.github.io/research_materials/2018/non_separable_penalty}{Dealing with a non-separable penalty}
}
\author{
Junkyu Park
}
