% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgf.R
\name{histo}
\alias{histo}
\title{histo}
\usage{
histo(x, dat, bins)
}
\arguments{
\item{x}{A number, or vector of numbers.}

\item{dat}{A vector of numbers.}

\item{bins}{A vector of numbers where \code{min(bins) <= min(dat)} and
\code{max(bins) > max(dat)}.}
}
\description{
Given one-dimensional data \code{dat} and bins \code{bins}, evaluate the
histogram function at \code{x}.
}
\examples{
# Example: dX in csum_N
# Evaluate the histogram of 10000 samples of S = sum of iid X_i ~ dX
# at 0, 1, ..., M - 1 for some integer M, where the endpoints of bins are
# set to be seq(-.5, M - .5, by = 1).
result_X <- csum_N(dX, support = 0:5, lambda = 3)
M <- length(result_X)
samples_N <- rpmf(10000, dpmf, 0:(M - 1), result_X)
histo(0:(M - 1), samples_N, seq(-.5, M - .5, by = 1))
}
\seealso{
\code{\link{csum_N}}
  \code{\link{dpmf}}
  \code{\link{rpmf}}
}
\author{
Junkyu Park
}
