% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_imputation.R
\name{impute_em}
\alias{impute_em}
\title{impute_em}
\usage{
impute_em(X, max_iter = 3000, eps = 1e-08)
}
\arguments{
\item{X}{A data or a matrix with or without \code{NA}'s.}

\item{max_iter}{A natural number; the number of iterations.
\code{3000} by default.}

\item{eps}{A small positive number; 1e-08 by default.}
}
\value{
A list:
  \itemize{
    \item \code{mu}: The resulting mean vector estimate. If \code{X} (input)
      has no missing components, then this is the MLE.
    \item \code{Sigma}: The resulting variance matrix estimate. If \code{X}
      (input) has no missing components, then this is the MLE.
    \item \code{X_imputed}: An imputed data.
    \item \code{C}: \code{!is.na(X)}, where \code{X} is the input.
    \item \code{iter}: The number of iterations.
  }
}
\description{
Impute missing components in \code{X} using the EM algorithm, assuming
each row is a sample of multivariate normal distribution.
}
\examples{
set.seed(1024)
mu <- c(1, 2, 6)
Sigma <- matrix(c(118, 62, 44, 62, 49, 17, 44, 17, 21), nrow = 3)
n <- 400
X_truth <- MASS::mvrnorm(n = n, mu = mu, Sigma = Sigma)
result <- simulate_na(X_truth, na_rate = .4)
result_imputed <- impute_em(result$X)
}
\seealso{
\code{\link{simulate_na}},
  \href{https://joon3216.github.io/research_materials/2019/em_imputation}{Imputing missing data using EM algorithm}
}
\author{
Junkyu Park
}
