% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{plot_roc}
\alias{plot_roc}
\title{plot_roc}
\usage{
plot_roc(dat, fmlr)
}
\arguments{
\item{dat}{A data whose response is binary (0 and 1).}

\item{fmlr}{A formula for the logistic regression with logit link.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the ROC curve of the logistic regression with logit link.
}
\examples{
# data(nodal, package = 'SMPracticals')
plot_roc(nodal, r ~ stage + xray + acid)
}
\seealso{
\code{\link{binarize_binom}},
  \code{\link{binarize_pois}}
  \code{\link{change_form}}
  \code{\link{CI_auc}}
}
\author{
Junkyu Park
}
