% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{random_word}
\alias{random_word}
\title{random_word}
\usage{
random_word(n)
}
\arguments{
\item{n}{A natural number. This is the number of characters in a generated
word.}
}
\value{
A character having \code{n} characters.
}
\description{
Generate a random combination of English alphabets of length \code{n}.
}
\examples{
set.seed(1024)
random_word(5)
random_word(10)
}
\seealso{
\code{\link{binarize_binom}},
  \code{\link{binarize_pois}}
}
\author{
Junkyu Park
}
