% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funpark_data.R
\docType{data}
\name{rich}
\alias{rich}
\title{rich dataset}
\format{A data.table with 10 rows and 3 variables:
\describe{
    \item{gender}{gender, either \code{M}ale or \code{F}emale}
    \item{criminal_record}{indicating whether a person has a criminal record (\code{yes}) or not (\code{no})}
    \item{is_rich}{integer column, \code{0} if not rich and \code{1} if rich}
}}
\usage{
data(rich)
}
\description{
A sample dataset to test functions. This dataset has a binary response.
}
\examples{
data(rich)
change_form(
    rich,
    from = 'binary', to = 'binomial',
    old_response = 'is_rich',
    new_response = c('rich', 'not_rich')
)
}
\keyword{datasets}
