% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funpark_data.R
\docType{data}
\name{rich_binom}
\alias{rich_binom}
\title{rich_binom dataset}
\format{A data.table with 4 rows and 4 variables:
\describe{
    \item{gender}{gender, either \code{M}ale or \code{F}emale}
    \item{criminal_record}{indicating whether a person has a criminal record (\code{yes}) or not (\code{no})}
    \item{rich}{the number of rich people that satisfies \code{(gender, criminal_record)} combination}
    \item{not_rich}{the number of non-rich people that satisfies the same combination as in \code{rich}}
}}
\usage{
data(rich_binom)
}
\description{
A sample dataset to test functions. This dataset has binomial responses.
}
\examples{
data(rich_binom)
binarize_binom(rich_binom, c('rich', 'not_rich'), 'is_rich')
}
\keyword{datasets}
