% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funpark_data.R
\docType{data}
\name{rich_pois}
\alias{rich_pois}
\title{rich_pois dataset}
\format{A data.table with 8 rows and 4 variables:
\describe{
    \item{gender}{gender, either \code{M}ale or \code{F}emale}
    \item{criminal_record}{indicating whether a person has a criminal record (\code{yes}) or not (\code{no})}
    \item{is_rich}{factors, \code{rich} if a person is rich and \code{not_rich} otherwise}
    \item{count}{the number of people that satisfies \code{(gender, criminal_record, is_rich)} combination}
}}
\usage{
data(rich_pois)
}
\description{
A sample dataset to test functions. This dataset has a count response.
}
\examples{
data(rich_pois)
binarize_pois(rich_pois, 'count')
}
\keyword{datasets}
