% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgf.R
\name{rpmf}
\alias{rpmf}
\title{rpmf}
\usage{
rpmf(n, pmf, support, ...)
}
\arguments{
\item{n}{An integer >= 1.}

\item{pmf}{A probability mass function which takes elements of an ambient space of pmf's support (i.e. an input that makes pmf return a probability) in its first argument. This pmf has a finite \code{support}, and must be able to take an array of elements and return an array of probabilities. Each output of this \code{pmf} must fall in [0, 1], and every conceivable output, including 0, must sum up to 1.}

\item{support}{A vector which consists of the elements of the support of \code{pmf}, i.e. \code{sum(pmf(support, ...)) == 1 && !(0 \%in\% pmf(support, ...))}}

\item{\dots}{An additional argument of \code{pmf}}
}
\description{
Draw random samples from a probability mass function using the \href{https://en.wikipedia.org/wiki/Inverse_transform_sampling}{inverse transform sampling}.
}
\examples{
# Example: dfruit in dpmf example
# Generate 12 random samples from dfruit
rpmf(12, dfruit, c('apple', 'orange', 'neither'))
}
\seealso{
\code{\link{dpmf}}
}
\author{
Junkyu Park
}
