% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_imputation.R
\name{simulate_na}
\alias{simulate_na}
\title{simulate_na}
\usage{
simulate_na(X_complete, na_rate)
}
\arguments{
\item{X_complete}{A data or a matrix with no \code{NA}'s.}

\item{na_rate}{A number in (0, 1).}
}
\value{
A list:
  \itemize{
    \item \code{X}: A generated data with missing components.
    \item \code{C}: \code{!is.na(X)}
    \item \code{na_rate}: A specified \code{na_rate} when running the
      function.
    \item \code{na_rate_actual}: The actual proportion of missing components
      after running the function.
  }
}
\description{
Randomly replace some components in \code{X_complete} with \code{NA} based
on \code{na_rate} specified.
}
\examples{
set.seed(1024)
mu <- c(1, 2, 6)
Sigma <- matrix(c(118, 62, 44, 62, 49, 17, 44, 17, 21), nrow = 3)
n <- 400
X_truth <- MASS::mvrnorm(n = n, mu = mu, Sigma = Sigma)
result <- simulate_na(X_truth, na_rate = .4)
}
\seealso{
\code{\link{impute_em}},
  \href{https://joon3216.github.io/research_materials/2019/em_imputation}{Imputing missing data using EM algorithm}
}
\author{
Junkyu Park
}
